% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{add one to each element of a vector}
\usage{
add_one(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The input numeric vector, with 1 added to each element.
}
\description{
add one to each element of a vector
}
\examples{
add_one(1:10)

}
