% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meancpp}
\alias{meancpp}
\title{Example 1 Vector Input Scalar Output}
\usage{
meancpp(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
y A double as mean
}
\description{
Calculate mean of a numeric vector
}

