% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdistcpp}
\alias{pdistcpp}
\title{Example 2 Vector Input Vector Output}
\usage{
pdistcpp(x, ys)
}
\arguments{
\item{x}{A numeric number}
}
\value{
y A numeric vector
}
\description{
Calculate Euclidean distance
}

