\name{testrcpp-package}
\alias{testrcpp-package}
\alias{testrcpp}
\docType{package}
\title{
\packageTitle{testrcpp}
}
\description{
\packageDescription{testrcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testrcpp}
\packageIndices{testrcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{testrcpp}

Maintainer: \packageMaintainer{testrcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
