% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{Function to get the double of the value}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{A single value.}
}
\value{
A single value, double of x
}
\description{
Calculating the double of x, i.e. x times 2
}

